调整帮助信息和新增查看备份文件的命令

+++ shell/xtrab.sh
@@ -8,6 +8,22 @@ umask 077
 export TZ=Asia/Shanghai
 bak_day=$(date +"%Y%m%d-%H%M%S")
 
+_usage() {
+    echo "In Docker Container"
+    echo "Usage: $0 backup"
+    echo "     : backup  to ./backup/%Y%m%d-%H%M%S.tar.gz"
+    echo
+    echo "Usage: $0 restore [latest|20150625-123456.tar.gz]"
+    echo "     : restore the backup of latest or 20150625-123456 "
+    echo
+    echo "Usage: $0 delete  [7|15|30]"
+    echo "     : delete  the backup of [7|15|30] day ago"
+    echo "Usage: $0 list"
+    echo "     : ls -l ./backup/*tar.gz"
+    echo
+    exit 127;
+}
+
 temp_dir() {
   [ -d "./backup/temp" ] && rm -rf ./backup/temp
   mkdir ./backup/temp
@@ -42,10 +58,6 @@ _wait_sock() {
   done
 }
 restore() {
-  supervisorctl stop mysqld
-
-  [ -S "./logs/mysql.sock" ] && sleep 2
-  [ -S "./logs/mysql.sock" ] && { echo "stop mysqld error?"; exit 127; }
 
   temp_dir
   local backup_file="$1"
@@ -53,18 +65,26 @@ restore() {
     backup_file="./backup/latest-backup.tar.gz"
   else
     backup_file="./backup/$backup_file"
-    [ -f $backup_file ] || { echo " No backup file"; exit 127; }
+    [ -f $backup_file ] || { echo " No backup file...?";  _usage; }
   fi
 
+  # 停mysql
+  supervisorctl stop mysqld
+
+  [ -S "./logs/mysql.sock" ] && sleep 2
+  [ -S "./logs/mysql.sock" ] && { echo "stop mysqld error?"; exit 127; }
+
+  # 准备之前备份的数据
   tar xfi $backup_file -C  ./backup/temp/
   innobackupex --apply-log ./backup/temp/
 
-  # 移动现有data 和log 下的文件至备份目录
+  # 移动正式data 和log 下的文件至备份目录
   mkdir -v ./backup/data-${bak_day} ./backup/log-${bak_day}
   mv ./data/*      ./backup/data-${bak_day}
   mv ./data/.xtrab ./backup/data-${bak_day}
   mv ./log/*       ./backup/log-${bak_day}
 
+  # 将备份数据复制回正式目录
   innobackupex --defaults-file="/etc/mysql/my.cnf" \
                --copy-back ./backup/temp/
 
@@ -82,6 +102,7 @@ restore() {
 
 }
 
+# 删除备份文件
 delete(){
   local mtime="$2"
 
@@ -90,26 +111,33 @@ delete(){
    -exec rm -fv {} \;
 }
 
-case "$1" in
+# 列出备份文件
+_list() {
+  pushd ./backup >/dev/null
+  ls -l *tar.gz
+}
+
+prog=$(basename $0)
+if echo "$prog" | egrep -q "^(backup|restore|list)$" ;then
+  action="$prog"
+else
+  action="$1"
+fi
+
+case "$action" in
   backup)
     backup
-    ;;
+  ;;
   restore)
     restore $2
-    ;;
+  ;;
   delete)
     delete $2
-    ;;
+  ;;
+  list)
+    _list
+  ;;
   *)
-    echo "In Docker Container"
-    echo "Usage: $0 backup"
-    echo "     : backup  to ./backup/%Y%m%d-%H%M%S.tar.gz"
-    echo
-    echo "Usage: $0 restore [latest|20150625-123456.tar.gz]"
-    echo "     : restore the backup of latest or 20150625-123456 "
-    echo
-    echo "Usage: $0 delete  [7|15|30]"
-    echo "     : delete  the backup of [7|15|30] day ago"
-    echo
+    _usage
   ;;
 esac
